Reting how we do aggregation.
I wanted to be able to specify the data type contained in the vector
accumulated by `collecting`.  I want able to make that work using
functions so I introdiced aggregators.  These are structs that are
subptypes of `Aggregator`.  The two implemented here are `Counter` abd
`Collector{T}`.  Othere are possible.

all tests were converted to the new paridigm and some documentation
added.